Updated orders and users api.go
