+++ golang/Assignments/ecomapp/src/users/go.mod
@@ -1,0 +1,8 @@
+module ecomapp/src/users
+
+go 1.16
+
+require (
+	github.com/gorilla/mux v1.8.0
+	github.com/lib/pq v1.10.0
+)
