--- golang/Assignments/ecomapp/users/api.go
@@ -1,86 +1,0 @@
-package main
-
-import (
-	"database/sql"
-	"fmt"
-	"log"
-	"net/http"
-	"os"
-
-	_ "github.com/lib/pq"
-)
-
-var db *sql.DB
-
-const (
-	dbhost = "DBHOST"
-	dbport = "DBPORT"
-	dbuser = "DBUSER"
-	dbpass = "DBPASS"
-	dbname = "DBNAME"
-)
-
-func initDb() {
-	config := dbConfig()
-	var err error
-	psqlInfo := fmt.Sprintf("host=%s port=%s user=%s "+
-		"password=%s dbname=%s sslmode=disable",
-		config[dbhost], config[dbport],
-		config[dbuser], config[dbpass], config[dbname])
-
-	db, err = sql.Open("postgres", psqlInfo)
-	if err != nil {
-		panic(err)
-	}
-	err = db.Ping()
-	if err != nil {
-		panic(err)
-	}
-	userTableQuery := `CREATE TABLE IF NOT EXISTS users(user_id SERIAL PRIMARY KEY, user_name char(50))`
-	_, err = db.Exec(userTableQuery)
-	if err != nil {
-		panic(err)
-	}
-	fmt.Println("DB Initialized Successfully!")
-}
-
-func dbConfig() map[string]string {
-	conf := make(map[string]string)
-	host, ok := os.LookupEnv(dbhost)
-	if !ok {
-		panic("DBHOST environment variable required but not set")
-	}
-	port, ok := os.LookupEnv(dbport)
-	if !ok {
-		panic("DBPORT environment variable required but not set")
-	}
-	user, ok := os.LookupEnv(dbuser)
-	if !ok {
-		panic("DBUSER environment variable required but not set")
-	}
-	password, ok := os.LookupEnv(dbpass)
-	if !ok {
-		panic("DBPASS environment variable required but not set")
-	}
-	name, ok := os.LookupEnv(dbname)
-	if !ok {
-		panic("DBNAME environment variable required but not set")
-	}
-	conf[dbhost] = host
-	conf[dbport] = port
-	conf[dbuser] = user
-	conf[dbpass] = password
-	conf[dbname] = name
-	return conf
-}
-
-func main() {
-	initDb()
-	defer db.Close()
-	http.HandleFunc("/api/users", userHandler)
-	log.Fatal(http.ListenAndServe("localhost:8000", nil))
-}
-
-func userHandler(w http.ResponseWriter, r *http.Request) {
-	fmt.Fprintf(w, "User API")
-}
